Flytt valideringsfunksjoner i BehandlingsresultatSteg.kt inn i egen service
