+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/behandlingsresultat/BehandlingsresultatValideringUtils.kt
@@ -19,16 +19,29 @@ import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandlingsresultat.FORTSA
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandlingsresultat.FORTSATT_OPPHØRT
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandlingsresultat.IKKE_VURDERT
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandlingsresultat.OPPHØRT
+import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ba.sak.kjerne.beregning.domene.tilTidslinjerPerAktørOgType
 import no.nav.familie.ba.sak.kjerne.eøs.felles.util.MAX_MÅNED
 import no.nav.familie.ba.sak.kjerne.eøs.felles.util.MIN_MÅNED
+import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.Kompetanse
+import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.KompetanseResultat
+import no.nav.familie.ba.sak.kjerne.forrigebehandling.EndringIUtbetalingUtil
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import no.nav.familie.ba.sak.kjerne.tidslinje.transformasjon.beskjærTilOgMed
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårsvurderingForskyvningUtils.lagForskjøvetTidslinjeForOppfylteVilkår
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.PersonResultat
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.UtdypendeVilkårsvurdering.BOSATT_I_FINNMARK_NORD_TROMS
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.UtdypendeVilkårsvurdering.DELT_BOSTED
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.UtdypendeVilkårsvurdering.DELT_BOSTED_SKAL_IKKE_DELES
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkår
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkår.BOSATT_I_RIKET
+import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkårsvurdering
 import no.nav.familie.tidslinje.Tidslinje
+import no.nav.familie.tidslinje.utvidelser.kombinerMed
 import no.nav.familie.tidslinje.utvidelser.outerJoin
+import no.nav.familie.tidslinje.utvidelser.tilPerioder
 import no.nav.familie.tidslinje.utvidelser.tilPerioderIkkeNull
 import java.time.YearMonth
 
@@ -151,6 +164,133 @@ object BehandlingsresultatValideringUtils {
                 }
             }
         }
+    }
+
+    fun validerKompetanse(
+        kompetanser: Collection<Kompetanse>,
+    ) {
+        kompetanser.forEach { kompetanse ->
+            val erNorgeSekundærland = kompetanse.resultat == KompetanseResultat.NORGE_ER_SEKUNDÆRLAND
+            if (erNorgeSekundærland && setOf(kompetanse.søkersAktivitetsland, kompetanse.annenForeldersAktivitetsland, kompetanse.barnetsBostedsland).all { it == "NO" }) {
+                throw FunksjonellFeil("Dersom Norge er sekundærland, må søkers aktivitetsland, annen forelders aktivitetsland eller barnets bostedsland være satt til noe annet enn Norge")
+            }
+        }
+    }
+
+    fun validerFinnmarkstilleggBehandling(
+        andelerNåværendeBehandling: Collection<AndelTilkjentYtelse>,
+        andelerForrigeBehandling: Collection<AndelTilkjentYtelse>,
+        vilkårsvurdering: Vilkårsvurdering,
+        inneværendeMåned: YearMonth,
+    ) {
+        validerAtDetIkkeHarVærtEndringIUtbetalingUtenomYtelseType(
+            andelerNåværendeBehandling = andelerNåværendeBehandling,
+            andelerForrigeBehandling = andelerForrigeBehandling,
+            ytelseType = YtelseType.FINNMARKSTILLEGG,
+            behandlingÅrsak = BehandlingÅrsak.FINNMARKSTILLEGG,
+        )
+
+        validerAtIngenAndelerMedYtelseTypeErInnvilgetFramITid(
+            andeler = andelerNåværendeBehandling,
+            ytelseType = YtelseType.FINNMARKSTILLEGG,
+            inneværendeMåned = inneværendeMåned,
+        )
+
+        validerAtDetIkkeFinnesDeltBostedForBarnSomIkkeBorMedSøkerIFinnmark(vilkårsvurdering)
+    }
+
+    fun validerSvalbardtilleggBehandling(
+        andelerNåværendeBehandling: Collection<AndelTilkjentYtelse>,
+        andelerForrigeBehandling: Collection<AndelTilkjentYtelse>,
+        inneværendeMåned: YearMonth,
+    ) {
+        validerAtDetIkkeHarVærtEndringIUtbetalingUtenomYtelseType(
+            andelerNåværendeBehandling = andelerNåværendeBehandling,
+            andelerForrigeBehandling = andelerForrigeBehandling,
+            ytelseType = YtelseType.SVALBARDTILLEGG,
+            behandlingÅrsak = BehandlingÅrsak.SVALBARDTILLEGG,
+        )
+
+        validerAtIngenAndelerMedYtelseTypeErInnvilgetFramITid(
+            andeler = andelerNåværendeBehandling,
+            ytelseType = YtelseType.SVALBARDTILLEGG,
+            inneværendeMåned = inneværendeMåned,
+        )
+    }
+
+    private fun validerAtDetIkkeFinnesDeltBostedForBarnSomIkkeBorMedSøkerIFinnmark(vilkårsvurdering: Vilkårsvurdering) {
+        val søkersPersonResultat = vilkårsvurdering.personResultater.find { it.erSøkersResultater() } ?: return
+
+        val søkerBosattIFinnmarkTidslinje =
+            søkersPersonResultat.vilkårResultater
+                .filter { it.vilkårType == BOSATT_I_RIKET && BOSATT_I_FINNMARK_NORD_TROMS in it.utdypendeVilkårsvurderinger }
+                .lagForskjøvetTidslinjeForOppfylteVilkår(BOSATT_I_RIKET)
+
+        vilkårsvurdering
+            .personResultater
+            .filterNot { it.erSøkersResultater() }
+            .forEach { personResultat ->
+                val barnBosattIFinnmarkTidslinje =
+                    personResultat.vilkårResultater
+                        .filter { it.vilkårType == BOSATT_I_RIKET && BOSATT_I_FINNMARK_NORD_TROMS in it.utdypendeVilkårsvurderinger }
+                        .lagForskjøvetTidslinjeForOppfylteVilkår(BOSATT_I_RIKET)
+
+                val barnDeltBostedTidslinje =
+                    personResultat.vilkårResultater
+                        .filter { it.vilkårType == Vilkår.BOR_MED_SØKER && (it.utdypendeVilkårsvurderinger.contains(DELT_BOSTED) || it.utdypendeVilkårsvurderinger.contains(DELT_BOSTED_SKAL_IKKE_DELES)) }
+                        .lagForskjøvetTidslinjeForOppfylteVilkår(Vilkår.BOR_MED_SØKER)
+
+                val finnesPerioderDerBarnMedDeltBostedIkkeBorSammenMedSøkerIFinnmark =
+                    søkerBosattIFinnmarkTidslinje.kombinerMed(barnBosattIFinnmarkTidslinje, barnDeltBostedTidslinje) { søkerBosattIFinnmark, barnBosattIFinnmark, barnDeltBosted ->
+                        søkerBosattIFinnmark != null && barnBosattIFinnmark == null && barnDeltBosted != null
+                    }
+
+                if (finnesPerioderDerBarnMedDeltBostedIkkeBorSammenMedSøkerIFinnmark.tilPerioder().any { it.verdi == true }) {
+                    throw Feil("Det finnes perioder der søker bor i finnmark samtidig som et barn med delt bosted ikke bor i finnmark. Disse sakene støtter vi ikke automatisk, og vi stanser derfor denne behandlingen.")
+                }
+            }
+    }
+
+    private fun validerAtDetIkkeHarVærtEndringIUtbetalingUtenomYtelseType(
+        andelerNåværendeBehandling: Collection<AndelTilkjentYtelse>,
+        andelerForrigeBehandling: Collection<AndelTilkjentYtelse>,
+        ytelseType: YtelseType,
+        behandlingÅrsak: BehandlingÅrsak,
+    ) {
+        val andelerUtenomYtelseTypeDenneBehandling = andelerNåværendeBehandling.filterNot { it.type == ytelseType }
+        val andelerUtenomYtelseTypeForrigeBehandling = andelerForrigeBehandling.filterNot { it.type == ytelseType }
+
+        val erEndringIUtbetaling =
+            EndringIUtbetalingUtil
+                .lagEndringIUtbetalingTidslinje(
+                    nåværendeAndeler = andelerUtenomYtelseTypeDenneBehandling,
+                    forrigeAndeler = andelerUtenomYtelseTypeForrigeBehandling,
+                ).tilPerioder()
+                .any { it.verdi == true }
+
+        if (erEndringIUtbetaling) {
+            throw Feil("Det er oppdaget forskjell i utbetaling utenom $ytelseType andeler. Dette kan ikke skje i en behandling der årsak er $behandlingÅrsak, og den automatiske kjøring stoppes derfor.")
+        }
+    }
+
+    private fun validerAtIngenAndelerMedYtelseTypeErInnvilgetFramITid(
+        andeler: Collection<AndelTilkjentYtelse>,
+        ytelseType: YtelseType,
+        inneværendeMåned: YearMonth,
+    ) {
+        val andelerMedYtelseType = andeler.filter { it.type == ytelseType }
+        val enMånedFramITid = inneværendeMåned.plusMonths(1)
+
+        andelerMedYtelseType
+            .groupBy { it.aktør }
+            .forEach { (_, andel) ->
+                val tidligsteAndelMedYtelseTypeForAktør = andel.minOfOrNull { it.stønadFom } ?: return@forEach
+
+                // TODO: Fiks valideringen når vi går live i oktober
+                if ((tidligsteAndelMedYtelseTypeForAktør > enMånedFramITid) && inneværendeMåned >= YearMonth.of(2025, 10)) {
+                    throw Feil("Det eksisterer $ytelseType andeler som først blir innvilget mer enn 1 måned fram i tid. Det er ikke mulig å innvilge disse enda, og behandlingen stoppes derfor.")
+                }
+            }
     }
 }
